Add compute pipeline classes
Queue family updates to add compute

Compute shaders work! (with lots of semaphore errors)

Refactor semaphores to be per swapchain rather than per frame

+++ src/asset/byte_writer.h
@@ -25,7 +25,7 @@ public:
 
     void WriteString(const std::string &str)
     {
-        uint32_t len = str.size() + 1;
+        uint32_t len = uint32_t(str.size()) + 1;
         char *tmp = new char[len];
         memset(tmp, 0, len);
         memcpy(tmp, str.c_str(), str.size());
